refactor(wasm): split device-state inspection into wasm/inspect.rs
Move the UI device-state accessors (board IO config/states, ADC/SPI/UART
device states, the SSD1306/ILI9341/PCD8544/SSD1680/UC8151D display
framebuffers + refresh generations, peripheral snapshot/list, iolink master
state) — 17 methods — out of the impl WasmSimulator god-block into
wasm/inspect.rs as a second #[wasm_bindgen] impl.

lib.rs 2392 -> 1659. Verified: wasm32 + host builds, clippy clean, cargo fmt.
